add spider assets in the game to test the new spider AI
