+++ Assets/Animation/Spider.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3193790d479784147bce5b3526a404e9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
