+++ Assets/Animation/Spider/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bd1e7db984af07d48821c22f50259844
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
