+++ Assets/Animation/Spider/Animations/Attack1.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b59a0a9f171f01d4aa83ea12b23b76f2
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
