+++ Assets/Animation/Spider/Animations/Run.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 43c0f719654248d44998a13dac0d6df4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
