+++ Assets/Animation/Spider/Animator.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 237d5c4b239caa746851deef4edec630
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
