+++ Assets/Animation/Spider/Animator/Enemy_Spider.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b9a0b43937c43a94b91266c4cdcfa16b
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
