+++ Assets/Prefabs/TreeLOD.prefab
@@ -165,6 +165,9 @@ PrefabInstance:
     - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 47a53fa538b5f7d4dba69b5594509e74, type: 3}
       insertIndex: -1
       addedObject: {fileID: -1032848759769105623}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 47a53fa538b5f7d4dba69b5594509e74, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 3885609574437580073}
   m_SourcePrefab: {fileID: 100100000, guid: 47a53fa538b5f7d4dba69b5594509e74, type: 3}
 --- !u!4 &8516515104824579812 stripped
 Transform:
@@ -241,3 +244,19 @@ CapsuleCollider:
   m_Height: 11.721039
   m_Direction: 1
   m_Center: {x: 0, y: 5.122467, z: 0}
+--- !u!208 &3885609574437580073
+NavMeshObstacle:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9041941559030910046}
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Shape: 1
+  m_Extents: {x: 1, y: 0.5, z: 1}
+  m_MoveThreshold: 0.1
+  m_Carve: 0
+  m_CarveOnlyStationary: 1
+  m_Center: {x: 0, y: 0, z: 0}
+  m_TimeToStationary: 0.5
