+++ Assets/Scenes/UIScene.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4217a04728a61144bb92482ea6b98b96
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
