+++ Assets/Scenes/UIScene.unity
@@ -319,6 +319,95 @@ MeshFilter:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 606608875}
   m_Mesh: {fileID: 3390558720134950976, guid: 2f59d7352614a0d41a464b77b57cb52d, type: 3}
+--- !u!1001 &610175881
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 148976109201488363, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_Size.x
+      value: 3.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 148976109201488363, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_Enabled
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 148976109201488363, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_Center.y
+      value: 0.9
+      objectReference: {fileID: 0}
+    - target: {fileID: 179937010355921882, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_Name
+      value: spider_black
+      objectReference: {fileID: 0}
+    - target: {fileID: 179937010355921882, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 1796145259731860449, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: enemyState
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3013755724190164062, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_Height
+      value: 1.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 3013755724190164062, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_Radius
+      value: 2
+      objectReference: {fileID: 0}
+    - target: {fileID: 3013755724190164062, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_Enabled
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 388.8894
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 71
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 287.3023
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 3372171602662781651, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 767a2296540c81e4194906bff1ec91ea, type: 3}
 --- !u!224 &645808395 stripped
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 9213338944092957178, guid: 016e8184f31654bada28e9c23b9a8b07, type: 3}
@@ -1087,7 +1176,6 @@ MonoBehaviour:
   _UIManager: {fileID: 4300147152962470403}
   encryption: 1
   maxLobbySize: 10
-  _stateManagerPrefab: {fileID: 8083658177136378179, guid: 0c5023e9d937d41ac9b0f7d05667380e, type: 3}
 --- !u!114 &1606576233
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -1224,11 +1312,54 @@ PrefabInstance:
       propertyPath: m_Name
       value: Environment
       objectReference: {fileID: 0}
+    - target: {fileID: 6375737767771200007, guid: d0fffed310d2f4993b1c553ea341674c, type: 3}
+      propertyPath: m_IsActive
+      value: 1
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
     m_RemovedGameObjects: []
     m_AddedGameObjects: []
-    m_AddedComponents: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 6375737767771200007, guid: d0fffed310d2f4993b1c553ea341674c, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 1771980031}
   m_SourcePrefab: {fileID: 100100000, guid: d0fffed310d2f4993b1c553ea341674c, type: 3}
+--- !u!1 &1771980030 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 6375737767771200007, guid: d0fffed310d2f4993b1c553ea341674c, type: 3}
+  m_PrefabInstance: {fileID: 1771980029}
+  m_PrefabAsset: {fileID: 0}
+--- !u!114 &1771980031
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1771980030}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 7a5ac11cc976e418e8d13136b07e1f52, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_AgentTypeID: 0
+  m_CollectObjects: 0
+  m_Size: {x: 10, y: 10, z: 10}
+  m_Center: {x: 0, y: 2, z: 0}
+  m_LayerMask:
+    serializedVersion: 2
+    m_Bits: 4294967295
+  m_UseGeometry: 0
+  m_DefaultArea: 0
+  m_GenerateLinks: 0
+  m_IgnoreNavMeshAgent: 1
+  m_IgnoreNavMeshObstacle: 1
+  m_OverrideTileSize: 0
+  m_TileSize: 256
+  m_OverrideVoxelSize: 0
+  m_VoxelSize: 0.11666667
+  m_MinRegionArea: 2
+  m_NavMeshData: {fileID: 23800000, guid: 92d3639286ed9e1479a959bc005d9728, type: 2}
+  m_BuildHeightMesh: 0
 --- !u!1001 &1841395270
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -2360,3 +2491,4 @@ SceneRoots:
   - {fileID: 4300147152962470402}
   - {fileID: 910107354}
   - {fileID: 9077078567108505936}
+  - {fileID: 610175881}
