+++ Assets/Scenes/UIScene/NavMesh-Environment.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 92d3639286ed9e1479a959bc005d9728
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
