--- Assets/Spider.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 807e8d8a4ee6ea84fb359443a7057cdf
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
