--- Assets/Spider/EnemyController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4520d21c613abb94a89bfaefec4b038a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
