--- Assets/Spider/Enemy_Spider.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: b9a0b43937c43a94b91266c4cdcfa16b
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
