--- Assets/Spider/spider_black.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 767a2296540c81e4194906bff1ec91ea
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
